+++ front/components/assistant/TryAssistant.tsx
